Partizan and Naginata Pass
Naginata retains the old peel, with both swing delay and standard divisor of 4. Because the naginata also gets long rend, which is chopping at 100% extra damage, and 20% extra damage on its cuts, it should be fine. Effectively no change, just that it doesn't get the partizan's increased effectiveness on peel.

Partizan gets a divisor of 3 on peel, with no delay, which makes it the best peeling weapon by far.
Defence was 6, for a super heavy defensive weapon, so I've given it 3 extra on wielded for 9, which brings it in line with a glaive that's unwielded. 6 on unwielded still.
Rend was kept the same, with the exception of misscost and delay removed / made standard reach rend values. Which brings it into line with other weapons using it, though being a stab, with less damage, as opposed to a chop.
It has also been given heavy weapon weight, which makes it better for draining stamina / breaking through parries, doubles sharpness/integrity damage in clashes, yet becomes easier to dodge.
As an aside, it also gets wielded bashing.

This should hopefully make the partizan something worth using, as opposed to the neglected side-choice that it has been forced into. A super heavy defensive weapon, intended to strip opponents of their armour and ability to fight effectively. It doesn't have the damage potential of the naginata, or the glaive's utility. But it should be playable and not genuinely crippling yourself as a result for using it over the aforementioned choices. A sidegrade with the potential to club, at best, and a solid weapon otherwise at worst. I hope.