Disables Corpse Qdel Spell, Enables Burial Rites
Field burials was added to improve performance of a server, in an area where we don't have concerns for that.
As such, it's vestigial, an odd element that lets you just immediately delete bodies without recourse, and can be used for farming mammon at an absurd rate.

This really didn't need to exist, and I've railed on it elsewhere.
Salt PR or whatever I guess. I don't care. It's questionable at best, and while it served its purpose on another community, as said, we don't need it.